prediction/events/requirements.txt
functions-framework==3.*
google-cloud-bigquery==3.*
pandas==2.*
requests==2.* 
python-dotenv==1.0.0